+++ lib/twitter_cldr/localized/localized_array.rb
@@ -3,6 +3,8 @@
 # Copyright 2012 Twitter, Inc
 # http://www.apache.org/licenses/LICENSE-2.0
 
+require 'yaml'
+
 module TwitterCldr
   module Localized
 
@@ -43,7 +45,7 @@ module TwitterCldr
       end
 
       def to_yaml(options = {})
-        TwitterCldr::Utils::YAML.dump(@base_obj, options)
+        ::YAML.dump(@base_obj, options)
       end
     end