+++ lib/twitter_cldr/localized/localized_hash.rb
@@ -3,12 +3,14 @@
 # Copyright 2012 Twitter, Inc
 # http://www.apache.org/licenses/LICENSE-2.0
 
+require 'yaml'
+
 module TwitterCldr
   module Localized
 
     class LocalizedHash < LocalizedObject
       def to_yaml(options = {})
-        TwitterCldr::Utils::YAML.dump(@base_obj, options)
+        ::YAML.dump(@base_obj, options)
       end
 
       def formatter_const