+++ lib/twitter_cldr/localized/localized_string.rb
@@ -3,6 +3,8 @@
 # Copyright 2012 Twitter, Inc
 # http://www.apache.org/licenses/LICENSE-2.0
 
+require 'yaml'
+
 module TwitterCldr
   module Localized
 
@@ -138,7 +140,7 @@ module TwitterCldr
       alias :each :each_char
 
       def to_yaml(options = {})
-        TwitterCldr::Utils::YAML.dump(@base_obj, options)
+        ::YAML.dump(@base_obj, options)
       end
 
       def to_bidi(options = {})