+++ lib/twitter_cldr/resources/aliases_importer.rb
@@ -5,6 +5,7 @@
 
 require 'fileutils'
 require 'nokogiri'
+require 'yaml'
 
 module TwitterCldr
   module Resources
@@ -23,9 +24,8 @@ module TwitterCldr
 
         File.open(output_file, 'w:utf-8') do |output|
           output.write(
-            TwitterCldr::Utils::YAML.dump(
-              TwitterCldr::Utils.deep_symbolize_keys(aliases: aliases),
-              use_natural_symbols: true
+            ::YAML.dump(
+              TwitterCldr::Utils.deep_symbolize_keys(aliases: aliases)
             )
           )
         end