+++ lib/twitter_cldr/resources/day_period_rules_importer.rb
@@ -5,6 +5,7 @@
 
 require 'nokogiri'
 require 'fileutils'
+require 'yaml'
 
 module TwitterCldr
   module Resources
@@ -29,9 +30,8 @@ module TwitterCldr
 
           File.open(output_file, 'w:utf-8') do |output|
             output.write(
-              TwitterCldr::Utils::YAML.dump(
-                TwitterCldr::Utils.deep_symbolize_keys(locale => rules),
-                use_natural_symbols: true
+              ::YAML.dump(
+                TwitterCldr::Utils.deep_symbolize_keys(locale => rules)
               )
             )
           end