+++ lib/twitter_cldr/resources/territories_importer.rb
@@ -4,6 +4,7 @@
 # http://www.apache.org/licenses/LICENSE-2.0
 
 require 'nokogiri'
+require 'yaml'
 
 module TwitterCldr
   module Resources
@@ -34,9 +35,8 @@ module TwitterCldr
 
         File.open(output_file, 'w:utf-8') do |output|
           output.write(
-            TwitterCldr::Utils::YAML.dump(
-              TwitterCldr::Utils.deep_symbolize_keys(locale => data),
-              use_natural_symbols: true
+            ::YAML.dump(
+              TwitterCldr::Utils.deep_symbolize_keys(locale => data)
             )
           )
         end