+++ lib/twitter_cldr/resources/timezones_importer.rb
@@ -8,6 +8,7 @@ require 'fileutils'
 require 'parallel'
 require 'etc'
 require 'set'
+require 'yaml'
 
 module TwitterCldr
   module Resources
@@ -47,9 +48,8 @@ module TwitterCldr
 
         File.open(output_file, 'w:utf-8') do |output|
           output.write(
-            TwitterCldr::Utils::YAML.dump(
-              TwitterCldr::Utils.deep_symbolize_keys(locale => data),
-              use_natural_symbols: true
+            ::YAML.dump(
+              TwitterCldr::Utils.deep_symbolize_keys(locale => data)
             )
           )
         end