+++ lib/twitter_cldr/resources/validity_data_importer.rb
@@ -5,6 +5,7 @@
 
 require 'fileutils'
 require 'nokogiri'
+require 'yaml'
 
 module TwitterCldr
   module Resources
@@ -20,9 +21,8 @@ module TwitterCldr
 
         File.open(output_file, 'w:utf-8') do |output|
           output.write(
-            TwitterCldr::Utils::YAML.dump(
-              TwitterCldr::Utils.deep_symbolize_keys(validity_data: validity_data),
-              use_natural_symbols: true
+            ::YAML.dump(
+              TwitterCldr::Utils.deep_symbolize_keys(validity_data: validity_data)
             )
           )
         end